Added a couple of features.
- Added the ability for clicking on the solution to automatically dive your currently location, and the ability to disable diving while solving.
- Added automatic detection of the oyster bed playing field.
- Added a fully automatic dive and swim feature.

Video demonstration: https://youtu.be/4LR39gyjIZg

Keep in mind, this was created against the old client and your mileage may vary in the new client as far as detection goes.  Also bear in mind that the detection can be very temperamental depending on the time of day and water reflection, as well as how closely you're zoomed in.  When it works good, it's great, when it doesn't...  Well, you can still input the puzzle manually and use the dive and solve feature with manual swimming.  The full auto dive and swim feature is dependent on auto detection catching the entire size of the puzzle.  If you don't detect at least 1 color (white/pinkish-white doesn't count) on all 4 sides, the grid dimensions may not be correct internally and I wouldn't recommend using the auto dive and swim feature.  Also, if you haven't guessed, that pinkish-white color is the most difficult to detect and account for due to the water reflections, so it is not part of computing the size and location of the grid.

Enjoy!